refactor(network): reorganize download
